AMP-10498
git-svn-id: file:///svn/branches/AMP_1_16_RELEASE@16060 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/digijava/module/aim/util/ActivityUtil.java
@@ -3834,7 +3834,7 @@ public static Long saveActivity(RecoverySaveParameters rsp) throws Exception {
   }
 
   public static ActivityAmounts getActivityAmmountIn(AmpActivity act,
-      String tocode,Long percent) throws Exception {
+      String tocode,Float percent) throws Exception {
     double tempProposed = 0;
     double tempActual = 0;
     double tempPlanned = 0;
